requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-mpd2==0.5.4
requests==2.9.1
Werkzeug==0.11.3
wheel==0.24.0
